Make canceling view non-atomic to fix 500 errors with job bursts
